+++ amundsen_application/static/js/components/Announcements/AnnouncementsList/styles.scss
