+++ amundsen_application/static/js/components/Announcements/constants.ts
